move context to approprate folders
