+++ src/components/dashboard/index.tsx
@@ -3,7 +3,7 @@ import { Link, Route, Switch } from 'react-router-dom';
 import { ExperienceEditor } from '../../components/experience-editor';
 import { GridPreview } from 'components/grid-preview';
 import { ModelPage } from 'pages/model.page';
-import { ModelProvider } from 'contexts/ModelContext';
+import { ModelProvider } from 'components/model/context';
 import { getModels } from 'services/strapi.service';
 import { ModelItemProvider } from 'components/model-item/context';
