+++ src/components/model/context/index.ts
@@ -1,0 +1,1 @@
+export { Context as ModelContext, Provider as ModelProvider, Consumer as ModelConsumer } from './model.context';
