--- src/contexts/ModelContext/index.ts
@@ -1,1 +1,0 @@
-export { Context as ModelContext, Provider as ModelProvider, Consumer as ModelConsumer } from './ModelContext';
